+++ CHANGELOG.md
@@ -7,9 +7,13 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+## [1.12.1] - 2018-08-22
+### Added
+- Product not found validation on ProductContextProvider
+
 ## [1.12.0] - 2018-08-21
 ### Added
-Canonical url for search context
+- Canonical url for search context
 
 ## [1.11.2] - 2018-08-20
 ### Changed